EpicHyperSketch

Documentation for EpicHyperSketch.

EpicHyperSketch.CountMinSketch

Base.getindex

Base.show

Base.size

EpicHyperSketch.cms_cols

EpicHyperSketch.cms_num_counters

EpicHyperSketch.cms_rows

EpicHyperSketch.config_size

EpicHyperSketch.CountMinSketch — Type

CountMinSketch(motif_size::Integer; delta, epsilon, use_cuda)

Create a CountMinSketch object.

motif_size: Number of features (motifs).

delta: Error probability.

epsilon: Error tolerance.

use_cuda: If true, use CUDA arrays.

Base.getindex — Method

getindex(cms::CountMinSketch, i, j)

Access the sketch matrix at position (i, j).

Base.show — Method

show(io::IO, cms::CountMinSketch)

Custom pretty printing for CountMinSketch objects.

Base.size — Method

size(cms::CountMinSketch)

Return the size of the sketch matrix.

EpicHyperSketch.cms_cols — Method

cms_cols(num_counters, rows)

Calculate the number of columns in the sketch given num_counters and rows.

EpicHyperSketch.cms_num_counters — Method

cms_num_counters(rows, epsilon)

Calculate the total number of counters in the sketch given rows and error tolerance epsilon.

Note: The counters simply means the total number of cells in the sketch matrix.

EpicHyperSketch.cms_rows — Method

cms_rows(delta)

Calculate the number of rows in the Count-Min Sketch given the error probability delta.

EpicHyperSketch.config_size — Method

config_size(num_features)

Return the configuration size for a given number of features.